Migrate MSVC tests to CodeBuild (#2583)
### Issues:
Resolves #CryptoAlg-3113 and #P274195896

### Description of changes: 
Migrate the following MSVC tests from GHA to CodeBuild:
- MSVC-2019
- MSVC-2022
- MSVC-SDE-64-bit

### Callouts:
The tests were running on Windows 2019 in GHA, but will be upgraded to
run in Windows 2022. `MSVC-SDE-64-bit` is the only test to be affected,
since some old CPUs don't support AVX instructions that Windows 2022 may
use. I updated `all_tests.go` to exclude those CPUs if Windows 2022 or
newer is detected.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.