Add support for EVP_PKEY_param_check (#2611)
### Issues:
Resolves https://github.com/ruby/openssl/issues/933

### Description of changes: 
Ironically, this was a discrepancy introduced due to our recent
endeavors for more OpenSSL compatibility. We introduced support for
`EVP_PKEY_check` and `EVP_PKEY_public_check` with bbf9ff6
and this was pulled into the latest release.

Unfortunately, ruby/openssl uses `EVP_PKEY_check` as a function support
check for `EVP_PKEY_param_check` and we happened to only implement
`EVP_PKEY_check` without `EVP_PKEY_param_check`. This PR implements
support for such to resolve the OpenSSL discrepancy.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.