Support for explicit curve parameter on EC public keys where parameters match supported curves (#2642)
### Description of changes: 

This feature is supported for private keys already, but EC public key
types did not support explicitly encoded curve parameters. This updates
the logic to support such types to use the same logic as EC private
keys, meaning that we will only accept explicit curve parameters that
match the named curve types we support. This does mean that subject
public keys in X.509 certificates, or that use ASN.1 encoding will now
support explicit parameters as well. This was a feature that OpenSSL
1.1.1 supported, but OpenSSL 3+ now will fail certificate validation on,
due to this being a non RFC conforming behavior and shouldn't happen in
publicly issued certificate by conforming certificate authorities. We
are adding this only for compatability with the former. I am still
investigating separately whether to disable / reject such keys in libssl
as to not open the door to much, and possibly whether we should add an
X509 verification flag to disable this behavior as well. Will follow
that up in a separate PR.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.