Fix benchmarking issues with FIPS main (#2655)
Turns out we never run against FIPS with main in our own benchmarking CI
tests. It took breaking the benchmarking canary for us to notice this.
This PR fixes the issue and adds the CI code path to ensure that this
doesn't break in the future. Also made some minor adjustments to fix
naming misalignments and a bug where we weren't running the right
benchmarking binary for the FIPS 2024 branch.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.