Increase SSLBuffer size to INT_MAX (#2673)
Curl is setting a buffer size larger than the maximum buffer length that
AWS-LC allows. `0x401e * 4` is inherently `65656`, which is larger than
the allowed `65535` and AWS-LC directly uses the largest size available
(`65535`) if the consuming application sets a value larger than that.
This limitation's due to a buffer size we inherited from upstream.

I've tested against the test that was failing for curl and the test
seems to pass now (for all instances I've ran it). Running `pytest -n
auto -v ./http/test_10_proxy.py::TestProxy::test_10_08_upload_seq_large
-rA` had been failing 5/10 times prior to this change and it passes
every time that I run it now.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.