Cherry-pick TLS transfer serialization changes to FIPS 3.x branch (#2672)
### Description of changes: 
* Cherry-pick's ONLY the TLS transfers changes from
[fa1c6c0ff647c5a998715e97e671acfd355ebad8](https://github.com/aws/aws-lc/commit/fa1c6c0ff647c5a998715e97e671acfd355ebad8)
which changes the boundary check condition.
* Cherry-pick's
[26da0177d6ee11defafb5624a0b387d275153222](https://github.com/aws/aws-lc/commit/26da0177d6ee11defafb5624a0b387d275153222)
* Cherry-pick's
[78ada214558fffe7be25455024780b5c666d5820](https://github.com/aws/aws-lc/commit/78ada214558fffe7be25455024780b5c666d5820)
* DOES NOT cherry-pick the TLS transfers changes in
[f7efc8bcad93365dbee9edbbbc1e4b3fea9e9cfb](https://github.com/aws/aws-lc/commit/f7efc8bcad93365dbee9edbbbc1e4b3fea9e9cfb)
which introduced serializing the `verify_result` field as an optional
field.

This is being done in order to make it more seamless to migrate from the
3.x to the future 4.x release and incorporate the serialization
improvements.

This changes does not impact FIPS module boundary.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Samuel Chiang <sachiang@amazon.com>