Make X509 CodeBuild webhook more resilient (#2680)
### Description of changes: 
Mainly reorganizes the project to have the x509-limbo and python
dependency already setup in a docker container (which will stay updates
by our periodic pipeline refreshment). We then use the image and pull in
the AWS-LC source and run the harness within that.

### Testing:
Verified the changes work with the X509 CodeBuild webhook project.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.