CodeBuild GitHub Actions Runner Project (#2704)
### Description of changes: 
Sets up a CodeBuild project to be able to leverage self-hosted GitHub
Action Runners in AWS CodeBuild.

### Testing:
Follow-up PRs will migrate various omnibus jobs and demonstrate this
working. For now this just stages the CDK code needed to get it setup so
we can phase roll it out.

A demonstration of this code in action can be found here:
https://github.com/skmcgrail/aws-lc/actions/runs/17870473934

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.