Fixes for android CI tests (#2713)
### Description of changes: 
Fixed failing Android CI tests.

### Call-outs:
The fix requires a Docker image update so devicefarm will still fail on
this PR. I have verified that the test passed in my personal account.
See test result for this
[commit](https://github.com/nhatnghiho/aws-lc/commits/a9d3dff8b25e82d3fe43d67951536ffee91e94e6)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Justin Smith <justsmth@amazon.com>