Implement workaround for FORTIFY_SOURCE warning with jitterentropy (#2728)
### Issues:
Resolves #2725

### Description of changes: 
We have received several reports where users are required to workaround
compilation issues with v1.60.0+ releases of AWS-LC where we
re-introduced jitterentropy into the library. This silences the
associated fortify source warnings if FORTIFY_SOURCE is enabled, as the
`-O0` optimization flag, required by jitterentropy's design, is not
compatible with it.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.