Add CI dimensions for legacy AVX512 flags (#2732)
### Issues:
`P299607876`

### Description of changes: 
We were previously only running the build, but not tests when options
MY_ASSEMBLER_IS_TOO_OLD_FOR_512AVX or MY_ASSEMBLER_IS_TOO_OLD_FOR_AVX
were enabled. This PR moves this testing to its own dimension and
enables running the tests.

Instead of the "run_tests" target, when GO is disabled,
"run_tests_minimal" is produced. Refactors to support this.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.