Add build-time option to opt-out of CPU Jitter Entropy (#2733)
Adds a build-time option DISABLE_USAGE_OF_CPU_JITTER_ENTROPY to opt-out of CPU Jitter Entropy. That is, with this option CPU Jitter Entropy will never be configured, the tree-DRBG will not be used. Furthermore, CPU Jitter Entropy will not even be included in any compilation units.

When using DISABLE_USAGE_OF_CPU_JITTER_ENTROPY we can no longer guarantee two entropy sources because only the Operating System type will be required to be available. rdrand and rndr will be used if supported in any given run-time, but it's not a guarantee.