CodeBuild Setup for GitHub Docker Image Builds (#2745)
### Description of changes: 
This is about one of three (or four) PRs that I need to parcel out to
roll out this change. This first introduces the necessary CodeBuild
project bits.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.