Cipher-stealing: no need for re-loading round keys; they're still in registers. (#2734)
In cipher-stealing at the end of the algorithm, the final encryption (or two decryptions) don't need to load the round keys again from memory, they are still kept in vector registers.
This also simplifies the HOL Light proof in reducing the assumptions about the memory reads of the round keys and keeping only the ones about the registers containing them.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Yan Peng <yppe@amazon.com>