AWS CodeBuild Fleets Setup (#2758)
### Description of changes: 
This adds a stack and configuration mechanism for setting up CodeBuild
fleets with auto-scaling capabilities. This will be used in subsequent
pull requests to migrate the graviton2 and graviton4 testing done with
the EC2 test framework, and allow for migration of Windows image builds.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.