Fix tpm2-tss CI (#2767)
### Issues:
Resolves `P316739151`

### Description of changes: 
Upstream updated their patch to use `1L` instead. The current patching
mechanism with `git apply` is also harder to debug, so I've changed it
to use `patch` instead.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.