AES-XTS on AArch64: Set w19 earlier before cipher-stealing of 1 block + tail. (#2785)
w19 was set to its fixed value 0x87 after the code jumps to the case of
cipher stealing of 1 block and a tail.
- This bug was introduced in
[v1.62](https://github.com/aws/aws-lc/releases/tag/v1.62.0) in "Move
udiv and sencond tweak calculations to when needed" by
[@nebeid](https://github.com/nebeid) in
[#2726](https://github.com/aws/aws-lc/pull/2726) where the setting of
w19 was moved past the branch to `.Lxts_enc_tail1x`.
- Also the 3 test vectors that tested 1 block + tail were not lucky
enough to exercise this issue which happens when the encrypted tweak,
when shifted left by 1, results in a bit of 1, which, in turn, results
in XORing the value 0x87 with the tweak.

Testing:
Added test vectors that failed with this bug (exercise the key/iv combination causing the failure) and now pass with the fix.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.