Use C++11 atomics to update session stats (#2786)
### Issues:
Addresses: #2776 
* This PR was tested/verified by [Willy Tarreau
(w@1wt.eu)](https://github.com/aws/aws-lc/pull/2786#issuecomment-3480923204)
who reported this issue.

### Description of changes: 
- `ssl_update_counter` uses atomic
[`fetch_add`](https://en.cppreference.com/w/cpp/atomic/atomic/fetch_add.html)
to increment when atomic operations are available.
- `ssl_read_counter` uses atomic
[`load`](https://en.cppreference.com/w/cpp/atomic/atomic/load.html) for
lock-free reads when available.
- Falls back to existing mutex-based approach when atomic operations are
unavailable or not guaranteed lock-free

### Call-outs:
The previous `ssl_read_counter` implementation was not taking a
reference to `counter`, so was acquiring the lock for no reason.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.