Use new images for fuzzing and x509 (#2804)
### Description of changes: 
Updates the fuzzing and x509 batch buildspecs to use the new container
images for fuzzing. Fuzzing and x509 remain as webhook initiated batch
builds, and there are no plans to migrate this to GitHub runner based
actions.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.