Disable old Windows jobs (#2812)
### Description of changes: 
Forgot to turn these off when I migrated the Windows jobs to the GitHub
Workflow file. For now this will be a stub echo job like all the others
in order to avoid failing the current webhook. This will be cleaned up
and removed once we migrate the FIPS branches over.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.