Setup OIDC for exchanging GitHub Token for AWS Credentials (#2819)
### Description of changes: 
This change configures AWS IAM with OpenID Connect to support assuming a
role in the AWS-LC team account by exchanging the GitHub token from a
workflow for temporary AWS credentials, this allows those credentials to
then be used to perform an assume role chain and use one of the
predefined policies depending on what the workflow's intention is. This
currently sets up three policies:

* A device farm role and policies which are analogous to the policy used
in the android CodeBuild webhook today.
* A docker image builder role and policies that capture the permissions
that are necessary for building and publishing our docker images to our
private registries.
* A standard GitHub actions role and policies that allow permissions to
pull images from our private docker registries and put metrics into our
CloudWatch namespace.

### Call-outs:
Nothing will change at the moment for the existing actions executing on
the AWS CodeBuild-managed GitHub Actions Runner environment. This is the
initial step to be able to scope down the service role policy that
implicitly gives workflows access to credentials for actions that aren't
always necessary. This PR sets up policies that I will migrate the
workflow files over to later, and then will revoke the permissions on
the service role.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.