Remove Docker Image build infrastructure from CodePipeline (#2822)
### Description of changes: 
File cleanups will come later, just getting this out of the pipeline so
the velocity of deployment can keep pace. These images are only relevant
to the FIPS branches at this point that still reference them, and the
Android device farm build. Both of those will be solved soon.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.