Adjust image-build-android concurrency group (#2848)
### Description of changes:
Concurrency groups can be problematic with similarly named branches,
especially with `pull_request_target`, this adjusts the concurrency
group so that it behaves consistently. Will make this adjustment to
other workflows as I move them to `pull_request_target` in the future.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.