Verify size of mlen in ML-DSA external mu mode (#2841)
### Issues:
Resolves #N/A

### Description of changes: 
When signing/verifying in ML-DSA the caller provides `mlen` the the
length of the message `m` that is being signed/verified. Currently, we
do no validation on the size of `mlen` when in pre-hash mode (called
"external mu" in ML-DSA). Since the pre-hash is the output of a SHAKE256
hash function, it is of fixed size `ML_DSA_CRHBYTES`. This adds
validation to check that, returning `-1` if not true to match the case
where the context `ctx` has length too large.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.