Add encap/decapKeyCheck support in ACVP (#2872)
ACVP has support for a new test vectors that test against the
encapsulation/decapsulation key checks defined in FIPS 203.
> Adds "encapsulationKeyCheck" and "decapsulationKeyCheck" as functions
for ML-KEM Encap/Decap FIPS203 to exercise an implementation's
capability to perform the Encapsulation Key Check in FIPS 203 Section
7.2 and the Decapsulation Key Check in FIPS 203 Section 7.3. These tests
are only included if the appropriate function is present in the
registration. They operate by providing a valid or invalid key and
expecting the IUT to return a true for a valid key or false for an
invalid key.

* https://github.com/usnistgov/ACVP-Server/releases

This add support in ACVP to run against the relevant
`crypto_kem_check_pk` and `crypto_kem_check_sk` functions that do the
checks for us. Also added the new `encapsulationKeyCheck` and
`decapsulationKeyCheck` test vectors in `ML-KEM.bz2`.

N/A

new `encapsulationKeyCheck` and `decapsulationKeyCheck` test vectors in
`ML-KEM.bz2`

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.