Prepare NetOS FIPS v1.29.1 release (#2879)
Prepare release of NetOS FIPS v1.29.1. The versioning scheme is a bit broken because the NetOS codepoint does not match any of the FIPSv releases. We cannot allocate v1.30.0 to this either, because that tag is taken by mainline AWS-LC.

Repurposing v1.29.1 is the least invasive we can do IMO. It's only meant to have an immutable reference to a codepoint intended for policy references and a release.