Support stdin for openssl rsa tool (#2899)
There are users that expect the stdin format when using the `openssl
rsa` tool. I also noticed a number of behavioral issues/differences
while working on this and comparing it to Openssl's version.
OpenSSL prioritizes the PKCS#8 SubjectPublicKeyInfo format first,
rather than the raw format. This is particularly problematic with
`stdin`, since `stdin` doesn't have a FILE rewind mechanism. This PR
aligns AWS-LC behavior with that.


### Call-outs:
N/A

### Testing:
New RSA comparison test

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Justin Smith <justsmth@amazon.com>