Iterate through all DNS entries in connect CLI
The `openssl sclient -connect` CLI command would only attempt to connect to the first resolved DNS entry. Failing that,
it considers the entire connection attempt failed. This change updates the command to iterate through the linked list of
entries until it can successfully connect. This is in-line with upstream [OpenSSL](https://github.com/openssl/openssl/blob/OpenSSL_1_1_1-stable/apps/s_socket.c#L93).

It's particularly useful in cases where the first address resolved is either unavailable or otherwise unusable. This can
be common when trying to resolve an address such as `localhost` and the OS returns the ipv6 loopback address before the
ipv4 version and your web server is only listening with ipv4.