Iterate through all DNS entries in connect CLI (#2906)
### Issues:
N/A

### Description of changes: 
The `openssl sclient -connect` CLI command would only attempt to connect
to the first resolved DNS entry. Failing that,
it considers the entire connection attempt failed. This change updates
the command to iterate through the linked list of
entries until it can successfully connect. This is in-line with upstream
[OpenSSL](https://github.com/openssl/openssl/blob/OpenSSL_1_1_1-stable/apps/s_socket.c#L93).

It's particularly useful in cases where the first address resolved is
either unavailable or otherwise unusable. This can
be common when trying to resolve an address such as `localhost` and the
OS returns the ipv6 loopback address before the
ipv4 version and your web server is only listening with ipv4.

### Call-outs:
I did not exhaustively consider other uses of this function beyond this
specific CLI tool. A glance check of the references suggests low usage
and it's still the right change to make overall.

### Testing:
- Tested this in a build container which resolves `localhost` as `[::1,
127.0.0.1]` with a web server that listens on ipv4 and asserted that the
CLI tool can connect successfully to that web server.

- CI

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.