Add openssl genpkey cli utility tool (#2907)
Internal consumers are depending on the `openssl genpkey` cli tool, so
this adds the necessary functionality to support that. This is mainly
used for generating a private key or key pair.
I've only added the necessary options for now which are `-algorithm`,
`-pkeyopt`, and `-out`. We can add more if needed, but there haven't
been more use cases discovered yet. I've also reused and cleaned up a
bit of the surrounding code while I was at it. Would like to refactor
more, but that would be a separate PR.

### Testing:
new unit tests. I haven't added any comparison tests, since it didn't
really make sense to compare generated private keys from two different
libraries.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.