Add stdin support for pkcs8 tool (#2915)
There are users that expect the stdin format when using the openssl
pkcs8 tool. I've removed some unnecessary cruft. There doesn't seem to
be an actual reason to cap the limit here, none of the other cli
implement this behavior and the relevant BIO functions will already fail
if the file loading becomes too excessive.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.